Use double float precision

+++ gmphd/birth_model.py
@@ -1,5 +1,5 @@
 import numpy as np
-
+from numpy import float64
 from gmphd.gaussian_component import GaussianComponent
 
 
@@ -10,11 +10,11 @@ class BirthModel:
 
     def __call__(self):
         birth_intensity = list()
-        mean = np.array([5., 5., 0.5, 0.5]).transpose()
+        mean = np.array([5., 5., 0.5, 0.5], dtype=float64).transpose()
         covariance = np.array([[1., 0., 0., 0.],
                                [0., 1., 0., 0.],
                                [0., 0., 1., 0.],
-                               [0., 0., 0., 1.]])
-        weight = 0.1
+                               [0., 0., 0., 1.]], dtype=float64)
+        weight = float64(0.1)
         birth_intensity.append(GaussianComponent(mean=mean, covariance=covariance, weight=weight))
         return birth_intensity
